+++ src/modules/Hooks.ts
@@ -49,6 +49,10 @@ export type HookArguments<T extends HookName> = HookDefinitions[T];
 
 export type HookName = keyof HookDefinitions;
 
+export type HookNameWithModifier = `${HookName}.${HookModifier}`;
+
+type HookModifier = 'once' | 'before' | 'replace';
+
 /** A generic hook handler. */
 export type HookHandler<T extends HookName> = (
 	/** Context about the current visit. */
@@ -69,6 +73,12 @@ export type HookDefaultHandler<T extends HookName> = (
 
 export type Handlers = {
 	[K in HookName]: HookHandler<K>[];
+};
+
+export type HookInitOptions = {
+	[K in HookName as K | `${K}.${HookModifier}`]: HookHandler<K>;
+} & {
+	[K in HookName as K | `${K}.${HookModifier}.${HookModifier}`]: HookHandler<K>;
 };
 
 /** Unregister a previously registered hook handler. */
@@ -516,7 +526,7 @@ export class Hooks {
 			handler = [{ id: 0, hook, handler: defaultHandler }];
 			if (replaced) {
 				const index = replace.length - 1;
-				const replacingHandler = replace[index].handler;
+				const { handler: replacingHandler, once } = replace[index];
 				const createDefaultHandler = (index: number): HookDefaultHandler<T> | undefined => {
 					const next = replace[index - 1];
 					if (next) {
@@ -527,9 +537,7 @@ export class Hooks {
 					}
 				};
 				const nestedDefaultHandler = createDefaultHandler(index);
-				handler = [
-					{ id: 0, hook, handler: replacingHandler, defaultHandler: nestedDefaultHandler }
-				];
+				handler = [{ id: 0, hook, once, handler: replacingHandler, defaultHandler: nestedDefaultHandler }]; // prettier-ignore
 			}
 		}
 
@@ -569,5 +577,15 @@ export class Hooks {
 		document.dispatchEvent(
 			new CustomEvent<HookEventDetail>(`swup:${hook}`, { detail, bubbles: true })
 		);
+	}
+
+	/**
+	 * Parse a hook name into the name and any modifiers.
+	 * @param hook Name of the hook.
+	 */
+	parseName(hook: HookName | HookNameWithModifier): [HookName, Partial<HookOptions>] {
+		const [name, ...modifiers] = hook.split('.');
+		const options = modifiers.reduce((acc, mod) => ({ ...acc, [mod]: true }), {});
+		return [name as HookName, options];
 	}
 }