+++ src/modules/plugins.ts
@@ -5,7 +5,7 @@ export type Plugin = {
 	isSwupPlugin: true;
 	/** Name of this plugin */
 	name: string;
-	/** Version of this plugin. Currently not in use, defined here for backward compatiblity. */
+	/** Version of this plugin. Currently not in use, defined here for backward compatibility. */
 	version?: string;
 	/** The swup instance that mounted this plugin */
 	swup?: Swup;
@@ -68,10 +68,9 @@ export function unuse(this: Swup, pluginOrName: Plugin | string) {
 
 /** Find a plugin by name or reference. */
 export function findPlugin(this: Swup, pluginOrName: Plugin | string) {
-	return this.plugins.find(
-		(plugin) =>
-			plugin === pluginOrName ||
-			plugin.name === pluginOrName ||
-			plugin.name === `Swup${String(pluginOrName)}`
-	);
+	return this.plugins.find((plugin) => {
+		return typeof pluginOrName === 'string'
+			? [`Swup${pluginOrName}`, pluginOrName].includes(plugin.name)
+			: plugin === pluginOrName;
+	});
 }