+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [4.8.3] - 2026-02-24
+
+- Remove `postinstall` script to prevent warnings when installing swup via pnpm
+
+## [4.8.2] - 2025-04-29
+
+- Fix `matchPath` to treat an empty array like an empty string
+
+## [4.8.1] - 2024-09-28
+
+- Fix usage of combined hook options `once` and `replace`
+
+## [4.8.0] - 2024-09-25
+
+- Add `hooks` option to allow registering all hooks at once
+- Add `visit.meta` key for custom data
+
 ## [4.7.0] - 2024-07-18
 
 - Inherit `data-swup-*` attributes from parent elements
@@ -330,6 +347,10 @@ See [upgrade instructions](https://swup.js.org/getting-started/upgrading-v3/) fo
 
 - Initial release
 
+[4.8.3]: https://github.com/swup/swup/releases/tag/4.8.3
+[4.8.2]: https://github.com/swup/swup/releases/tag/4.8.2
+[4.8.1]: https://github.com/swup/swup/releases/tag/4.8.1
+[4.8.0]: https://github.com/swup/swup/releases/tag/4.8.0
 [4.7.0]: https://github.com/swup/swup/releases/tag/4.7.0
 [4.6.1]: https://github.com/swup/swup/releases/tag/4.6.1
 [4.6.0]: https://github.com/swup/swup/releases/tag/4.6.0